+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/ExoPlayerEventListener.java
@@ -81,7 +81,6 @@ final class ExoPlayerEventListener implements Player.Listener {
         height = videoSize.width;
       }
     }
-
     events.onInitialized(width, height, exoPlayer.getDuration(), rotationCorrection);
   }