+++ packages/video_player/video_player_android/CHANGELOG.md
@@ -1,8 +1,11 @@
 ## 2.7.14
 
-* Changes the rotation correction calculation for Android API 21+ to use
-  that which is reported by the video's format instead of the unapplied
+* Changes the rotation correction calculation for Android API 29+ to use
+  the one that is reported by the video's format instead of the unapplied
   rotation degrees that Exoplayer does not report on Android API 21+.
+* Changes the rotation correction calculation for Android APIs 21-28 to 0
+  because the Impeller backend used on those API versions correctly rotates
+  the video being played automatically.
 
 ## 2.7.13